+++ lib/scss_lint/runner.rb
@@ -11,11 +11,11 @@ module SCSSLint
     def initialize(options = {})
       @lints = []
 
-      ignored_linters = LinterRegistry.
-        extract_linters_from(options.fetch(:ignored_linters, []))
+      excluded_linters = LinterRegistry.
+        extract_linters_from(options.fetch(:excluded_linters, []))
 
       @linters = LinterRegistry.linters.reject do |linter|
-        ignored_linters.include?(linter)
+        excluded_linters.include?(linter)
       end.map do |linter_class|
         linter_class.new
       end