+++ spec/cli_spec.rb
@@ -27,12 +27,13 @@ describe SCSSLint::CLI do
       end
     end
 
-    context 'when the ignore lints flag is set' do
-      let(:options) { ['-i', 'some_linter_name'] }
 
-      it 'sets the :ignored_linters option' do
+    context 'when the exclude linters flag is set' do
+      let(:options) { ['-x', 'SomeLinterName'] }
+
+      it 'sets the :excluded_linters option' do
         safe_parse
-        subject.options[:ignored_linters].should == ['some_linter_name']
+        subject.options[:excluded_linters].should == ['SomeLinterName']
       end
     end