+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Change `--ignore-linter` flag to use `CamelCase` linter names instead of
   `snake_case`
 * Removed `-x` alias for `--xml` flag
+* Change `-i`/`--ignore-linter` flags to `-x`/`--exclude-linter`
 
 0.7.1
 * Upgrade Sass gem dependency from 3.2.8 -> 3.2.9